backend/pyproject.toml
[tool.poetry]
name = "homeprice"
version = "0.1.0"
description = "backend app scoring price home"
authors = ["Brayan <codecsrayo@gmail.com>"]
license = "mit"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlmodel = "^0.0.22"
fastapi = "^0.114.1"
uvicorn = "^0.30.6"
pydantic = "^2.9.1"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"
faker = "^28.4.1"
alembic = "^1.13.2"
pydantic-settings = "^2.5.2"
asyncpg = "^0.29.0"
sqlalchemy-utils = "^0.41.2"
fastapi-pagination = "^0.12.27"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
